refactor(test-harness): one type per file — dissolve the render module
render.rs mixed five structs, four consts, six types' rendering impls, and
eight free functions. Each lands where it belongs:

- SuiteRow, Truncated, ListingSection, CompileView, SummaryTemplate each get
  their own module.
- The cell/line rendering impls move onto their types: DiffCounter::cell,
  SuiteStats::{row,*_cell}, CompileAggregate::cell, OutputVerdict::line,
  FailureVerdict::line.
- The consts become associated: Truncated::MAX_LISTED,
  CompileAggregate::SUITE_THRESHOLD_PERCENT, CompileView::PROJECT_THRESHOLD_PERCENT,
  SuiteStats::JITTER_MEDIAN_FLOOR_PERCENT.
- The free functions become associated: SuiteFailures::kinds, and the
  orchestration (health_lines, new_failure_bullets, baseline_rows,
  unpaired_runs_part, truncated_mode_list, and the render entry point) become
  SummaryTemplate associated functions.
